move commands into DriveTrain folder
