+++ src/main/java/frc/robot/commands/DriveTrainCommands/DriveStraightCommand.java
@@ -2,21 +2,18 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
-
-//import com.swervedrivespecialties.swervelib.SwerveModule;
+package frc.robot.commands.DriveTrainCommands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.DrivetrainSubsystem;
 
-public class sturdyBaseCommand extends CommandBase {
+public class DriveStraightCommand extends CommandBase {
   DrivetrainSubsystem m_drive;
-
-  /** Creates a new sturdyBaseCommand. */
-  public sturdyBaseCommand(DrivetrainSubsystem drive) {
-    m_drive = drive; 
-       // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements(drive);
+  /** Creates a new DriveStraightCommand. */
+  public DriveStraightCommand(DrivetrainSubsystem drive) {
+    m_drive = drive;
+    // Use addRequirements() here to declare subsystem dependencies.
+    addRequirements(m_drive);
   }
 
   // Called when the command is initially scheduled.
@@ -26,16 +23,14 @@ public class sturdyBaseCommand extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    m_drive.backLeftModule.set(0, -45);
-    m_drive.backRightModule.set(0, 45);
-    m_drive.frontRightModule.set(0, -45);
-    m_drive.frontLeftModule.set(0, 45);
-    
+    // m_drive.driveStraightSlow();
   }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    m_drive.stopModules();
+  }
 
   // Returns true when the command should end.
   @Override
